Crate kernel_loader

Linux kernel ELF file loader.

Modules

elf

Structs

Elf64

LoadedKernel

Information about a kernel loaded with the [load_kernel] function.

Enums

Error

Functions

load_cmdline

Writes the command line string to the given memory slice.

load_elf

Loads a kernel from a 32-bit or 64-bit ELF image into memory.

load_elf32

Loads a kernel from a 32-bit ELF image into memory.

load_elf64

Loads a kernel from a 64-bit ELF image into memory.

load_elf_for_class

read_elf

Reads the headers of an ELF32 or ELF64 object file. Returns ELF file and program headers, converted to ELF64 format. If required_ei_class is Some and the file’s ELF ei_class doesn’t match, an Err is returned.

read_elf_by_type

Reads the headers of an ELF32 or ELF64 object file. Returns ELF file and program headers, converted to ELF64 format. FileHeader and ProgramHeader are the ELF32 or ELF64 ehdr/phdr types to read from the file. Caller should check that file is a valid ELF file before calling this function.

Type Definitions

Result